week_6/requirements.txt
gradio
requests
google-generativeai
python-dotenv
mcp
numpy
pandas
python-dateutil
asyncio
pydantic